feat: split program data into new store
